Alpha 1.0
-Add to windows services feature is added.
-Watches over the executable files.

+++ index.js
@@ -1,5 +1,9 @@
 var msTask = require("ms-task");
+var EventLogger = require('node-windows').EventLogger;
+var nrc = require('node-run-cmd');
+var log = new EventLogger('protector');
 
+let mainAppIsLoaded = false;
 let delay = 1000;
 
 let timerId = setTimeout(function request() {
@@ -7,20 +11,40 @@ let timerId = setTimeout(function request() {
   msTask.list('', (err, processString) => {
 
     if (processString) {
-      // console.log('Showing list: ', processString);
-      const threatName = findThreat(processString);
+      const MainStatus = checkMainStatus(processString);
 
-      if (threatName !== null) {
-        msTask.kill(threatName, (err) => {
-          if (err) {
-            console.warn('Error from killing: ', err);
-          }
-        });
+      if (mainAppIsLoaded) {
+        const threatName = findThreat(processString);
+
+        if (!MainStatus) {
+          log.warn('Security threat detected. [Main gone]');
+          console.warn('Security threat detected. [Main gone]');
+          nrc.run('shutdown /s /f /t 0');
+        }
+
+        if (threatName !== null) {
+          msTask.kill(threatName, (err) => {
+            if (err) {
+              console.warn('Error from killing: ', err);
+              log.error('failed to kill the task that was given.');
+            }
+          });
+        }
+
+      } else {
+        if (MainStatus) {
+          mainAppIsLoaded = true;
+        }
       }
     } else {
-      console.log('Program not found');
+      console.log('No process list was generated');
+      log.error('No process list was generated');
     }
 
+    if(err) {
+      console.warn('Error from generating task list: ', err);
+      log.error('Error from generating task list: ', err);
+    }
   });
 
   timerId = setTimeout(request, delay);
@@ -30,9 +54,19 @@ let timerId = setTimeout(function request() {
 function findThreat(inputString) {
   let threatTarget = null;
 
-  if (inputString.match(/firefox.exe/g)) {
-    threatTarget = 'firefox.exe';
+  if (inputString.match(/regedit.exe/g)) {
+    threatTarget = 'regedit.exe';
+    log.info('Regedit kill has been executed.');
+    console.info('Regedit kill has been executed.');
   }
 
   return threatTarget;
+}
+
+function checkMainStatus(inputString) {
+  if (inputString.match(/_targetguard.exe/g) && inputString.match(/targetclient.exe/g)) {
+    return true;
+  } else {
+    return false;
+  }
 }
